day 03

+++ server/src/routes/memories.ts
@@ -3,8 +3,15 @@ import { prisma } from "../lib/prisma";
 import { z } from "zod";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/memories", async () => {
+  app.addHook("preHandler", async (request) => {
+    await request.jwtVerify();
+  });
+
+  app.get("/memories", async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: "asc",
       },
@@ -16,7 +23,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
       excerpt: content.substring(0, 155).concat("..."),
     }));
   });
-  app.get("/memories/:id", async (request) => {
+  app.get("/memories/:id", async (request, reply) => {
     const paramSchema = z.object({
       id: z.string().uuid(),
     });
@@ -28,6 +35,9 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       },
     });
+
+    if (!memory.isPublic && memory.userId !== request.user.sub)
+      return reply.status(401).send();
 
     return memory;
   });
@@ -46,14 +56,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: "47c8eb3c-5c8f-424a-b5e8-dc56f93b4abb",
+        userId: request.user.sub,
       },
     });
 
     return memory;
   });
 
-  app.put("/memories/:id", async (request) => {
+  app.put("/memories/:id", async (request, reply) => {
     const paramSchema = z.object({
       id: z.string().uuid(),
     });
@@ -68,7 +78,15 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, isPublic, coverUrl } = bodySchema.parse(request.body);
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== request.user.sub) return reply.status(401).send();
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -82,12 +100,18 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory;
   });
 
-  app.delete("/memories/:id", async (request) => {
+  app.delete("/memories/:id", async (request, reply) => {
     const paramSchema = z.object({
       id: z.string().uuid(),
     });
 
     const { id } = paramSchema.parse(request.params);
+
+    const memory = await prisma.memory.findFirstOrThrow({
+      where: { id },
+    });
+
+    if (memory.userId !== request.user.sub) return reply.status(401).send();
 
     await prisma.memory.delete({
       where: {
